+++ app/src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Register from "./components/RegisterComponent";  
 import Login from "./components/LoginComponent"
 import Welcome from "./components/WelcomeComponent"
-import Competition from "./components/CompetitionComponent";
+import Competition from "./pages/CompetitionComponent";
 import Perfil from "./components/PerfilComponent";
 import Rewards from "./components/RewardsComponent";
 import Missions from "./components/MissionsComponent";
