+++ app/src/components/ContentComponent.js
@@ -1,5 +1,6 @@
-
 import React, { useState, useEffect } from 'react';
+import Header from "./Header"; // Import Header component
+import Footer from "./Footer"; // Import Footer component
 import logo from '../img/Logo_Blue.png';
 import discord from '../img/discord.png';
 import instagram from '../img/instagram.png';
@@ -62,23 +63,9 @@ const Contenido = () => {
   }, [modalVisible]);
 
   return (
-    <div className="font-sans bg-blue-900 text-black m-0 p-0 min-h-screen flex flex-col"> {/* Add min-h-screen and flex classes */}
-    <header className="flex justify-between items-center px-4 py-4 md:px-6 bg-[#18538a]">
-<a href="/" className="w-[70px] h-[70px] mr-8">
-  <img src={oxygen} alt="OxygenGaming Logo" className="w-full h-full" />
-</a>
-<nav className="flex items-center gap-2">
-  <a href="/" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md hover:bg-white/10 bg-blue-600">Inicio</a>
-  <a href="/competiciones" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md hover:bg-white/10">Competiciones</a>
-  <a href="/content" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md hover:bg-white/10">Contenido</a>
-  <a href="/missions" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md hover:bg-white/10">Misiones</a>
-  <a href="/rewards" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md hover:bg-white/10">Recompensas</a>
-</nav>
-<div className="flex items-center gap-2 ml-auto">
-  <a href="/register" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md border-2 border-white bg-transparent">Registrarse</a>
-  <a href="/login" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md bg-blue-600 ml-4">Iniciar Sesión</a>
-</div>
-</header>
+    <div>
+      <Header /> {/* Add Header component */}
+      <div className="font-sans bg-blue-900 text-black m-0 p-0 min-h-screen flex flex-col"> {/* Add min-h-screen and flex classes */}
     <div className="container p-5 flex justify-center items-center mt-16"> {/* Change mt-12 to mt-16 to move title down */}
         <h1 className="text-white text-4xl font-bold ml-80">Contenido</h1> {/* Change ml-10 to ml-20 to move it further to the right */}
     </div>
@@ -236,8 +223,9 @@ const Contenido = () => {
 
     </footer>
 </div>
-
-);
+      <Footer /> {/* Add Footer component */}
+    </div>
+  );
 
 };
