+++ app/src/components/Header.js
@@ -16,13 +16,13 @@ const Header = () => {
             Inicio
           </a>
           <a
-            href="/competiciones"
+            href="/competition"
             className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md hover:bg-white/10"
           >
             Competiciones
           </a>
           <a
-            href="/reglamentos"
+            href="/content"
             className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md hover:bg-white/10"
           >
             Contenido
